+++ src/adv0/utils.h
@@ -3,7 +3,7 @@
  *
  * Miha Grčar 2022
  *
- * Various utils
+ * Various utils for Adventure
  */
 
 #ifndef __UTILS_H
@@ -13,7 +13,7 @@
 
 char *strcpy(char *destination, const char *source);
 void exit(int status);
-int atoi(char *str);
+UINT16 atoi(char *str);
 
 UINT8 *fread(char *path, UINT8 *out, UINT16 pos, UINT16 len);
